# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MailCue API",
    "description": "Realistic email testing server REST API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/domains": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "List All Domains",
        "description": "List all managed email domains. **Admin only.**",
        "operationId": "list_all_domains_api_v1_domains_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      },
      "post": {
        "tags": [
          "Domains"
        ],
        "summary": "Create Domain",
        "description": "Add a new email domain and generate DKIM keys. **Admin only.**",
        "operationId": "create_domain_api_v1_domains_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/domains/{name}": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Get Domain",
        "description": "Get domain details including required DNS records. **Admin only.**",
        "operationId": "get_domain_api_v1_domains__name__get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Domains"
        ],
        "summary": "Delete Domain",
        "description": "Remove a managed domain and clean up config. **Admin only.**",
        "operationId": "delete_domain_api_v1_domains__name__delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/domains/{name}/verify-dns": {
      "post": {
        "tags": [
          "Domains"
        ],
        "summary": "Verify Domain Dns",
        "description": "Run live DNS checks for a domain. **Admin only.**",
        "operationId": "verify_domain_dns_api_v1_domains__name__verify_dns_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DnsCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/domains/{name}/dns-state": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Get Dns State",
        "description": "Refresh DNS lookups + return per-record expected/published/drift state.\n\nRead-only: never flips the canonical ``*_verified`` booleans (use\n``POST /verify-dns`` for that).  Designed to be cheap and pollable from\nthe admin UI every 60s.  **Admin only.**",
        "operationId": "get_dns_state_api_v1_domains__name__dns_state_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainDnsStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DnsCheckResponse": {
        "properties": {
          "mx_verified": {
            "type": "boolean",
            "title": "Mx Verified"
          },
          "spf_verified": {
            "type": "boolean",
            "title": "Spf Verified"
          },
          "dkim_verified": {
            "type": "boolean",
            "title": "Dkim Verified"
          },
          "dmarc_verified": {
            "type": "boolean",
            "title": "Dmarc Verified"
          },
          "mta_sts_verified": {
            "type": "boolean",
            "title": "Mta Sts Verified",
            "default": false
          },
          "tls_rpt_verified": {
            "type": "boolean",
            "title": "Tls Rpt Verified",
            "default": false
          },
          "all_verified": {
            "type": "boolean",
            "title": "All Verified"
          },
          "dns_records": {
            "items": {
              "$ref": "#/components/schemas/DnsRecordInfo"
            },
            "type": "array",
            "title": "Dns Records"
          }
        },
        "type": "object",
        "required": [
          "mx_verified",
          "spf_verified",
          "dkim_verified",
          "dmarc_verified",
          "all_verified",
          "dns_records"
        ],
        "title": "DnsCheckResponse",
        "description": "Result of a live DNS verification check."
      },
      "DnsRecordInfo": {
        "properties": {
          "record_type": {
            "type": "string",
            "title": "Record Type"
          },
          "hostname": {
            "type": "string",
            "title": "Hostname"
          },
          "expected_value": {
            "type": "string",
            "title": "Expected Value"
          },
          "current_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Value"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "last_checked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checked At"
          },
          "last_verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified At"
          },
          "drift": {
            "type": "boolean",
            "title": "Drift",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "record_type",
          "hostname",
          "expected_value",
          "verified",
          "purpose"
        ],
        "title": "DnsRecordInfo",
        "description": "A single DNS record that needs to be configured."
      },
      "DomainCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "dkim_selector": {
            "type": "string",
            "title": "Dkim Selector",
            "default": "mail"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "DomainCreateRequest",
        "description": "Add a new email domain."
      },
      "DomainDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "dkim_selector": {
            "type": "string",
            "title": "Dkim Selector"
          },
          "mx_verified": {
            "type": "boolean",
            "title": "Mx Verified"
          },
          "spf_verified": {
            "type": "boolean",
            "title": "Spf Verified"
          },
          "dkim_verified": {
            "type": "boolean",
            "title": "Dkim Verified"
          },
          "dmarc_verified": {
            "type": "boolean",
            "title": "Dmarc Verified"
          },
          "mta_sts_verified": {
            "type": "boolean",
            "title": "Mta Sts Verified"
          },
          "tls_rpt_verified": {
            "type": "boolean",
            "title": "Tls Rpt Verified"
          },
          "last_dns_check": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Dns Check"
          },
          "all_verified": {
            "type": "boolean",
            "title": "All Verified"
          },
          "dns_records": {
            "items": {
              "$ref": "#/components/schemas/DnsRecordInfo"
            },
            "type": "array",
            "title": "Dns Records"
          },
          "dkim_public_key_txt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dkim Public Key Txt"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "is_active",
          "created_at",
          "dkim_selector",
          "mx_verified",
          "spf_verified",
          "dkim_verified",
          "dmarc_verified",
          "mta_sts_verified",
          "tls_rpt_verified",
          "last_dns_check",
          "all_verified",
          "dns_records"
        ],
        "title": "DomainDetailResponse",
        "description": "Extended domain info with DNS records and DKIM public key."
      },
      "DomainDnsStateResponse": {
        "properties": {
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/DnsRecordInfo"
            },
            "type": "array",
            "title": "Records"
          },
          "has_drift": {
            "type": "boolean",
            "title": "Has Drift"
          },
          "has_missing": {
            "type": "boolean",
            "title": "Has Missing"
          },
          "last_dns_check": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Dns Check"
          }
        },
        "type": "object",
        "required": [
          "domain",
          "records",
          "has_drift",
          "has_missing",
          "last_dns_check"
        ],
        "title": "DomainDnsStateResponse",
        "description": "Read-only snapshot of every published DNS record vs the expected value.\n\nDesigned to be polled cheaply from the admin UI to surface drift without\nflipping the canonical ``*_verified`` booleans (only ``POST /verify-dns``\ndoes that)."
      },
      "DomainListResponse": {
        "properties": {
          "domains": {
            "items": {
              "$ref": "#/components/schemas/DomainResponse"
            },
            "type": "array",
            "title": "Domains"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "domains",
          "total"
        ],
        "title": "DomainListResponse",
        "description": "List of managed domains."
      },
      "DomainResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "dkim_selector": {
            "type": "string",
            "title": "Dkim Selector"
          },
          "mx_verified": {
            "type": "boolean",
            "title": "Mx Verified"
          },
          "spf_verified": {
            "type": "boolean",
            "title": "Spf Verified"
          },
          "dkim_verified": {
            "type": "boolean",
            "title": "Dkim Verified"
          },
          "dmarc_verified": {
            "type": "boolean",
            "title": "Dmarc Verified"
          },
          "mta_sts_verified": {
            "type": "boolean",
            "title": "Mta Sts Verified"
          },
          "tls_rpt_verified": {
            "type": "boolean",
            "title": "Tls Rpt Verified"
          },
          "last_dns_check": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Dns Check"
          },
          "all_verified": {
            "type": "boolean",
            "title": "All Verified"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "is_active",
          "created_at",
          "dkim_selector",
          "mx_verified",
          "spf_verified",
          "dkim_verified",
          "dmarc_verified",
          "mta_sts_verified",
          "tls_rpt_verified",
          "last_dns_check",
          "all_verified"
        ],
        "title": "DomainResponse",
        "description": "Basic domain information."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}